requirements.txt
python-dotenv
slack_bolt
request
llama-index
rich
openai
etils[etree]
tqdm